+++ storage/file-torrent.go
@@ -46,30 +46,33 @@ func (me *fileTorrentImpl) setPieceCompletion(p int, complete bool) error {
 	return me.pieceCompletion().Set(me.pieceCompletionKey(p), complete)
 }
 
-func (me *fileTorrentImpl) markFileReleased(fileIndex int) error {
+func (me *fileTorrentImpl) validateFileIndex(fileIndex int) error {
 	if fileIndex < 0 || fileIndex >= len(me.files) {
 		return fmt.Errorf("file index %d out of range", fileIndex)
+	}
+	return nil
+}
+
+func (me *fileTorrentImpl) markFileReleased(fileIndex int) error {
+	if err := me.validateFileIndex(fileIndex); err != nil {
+		return err
 	}
 	f := me.file(fileIndex)
 	if err := me.writeReleasedBoundaryPieces(fileIndex, f); err != nil {
 		return err
 	}
 	f.mu.Lock()
+	if err := me.writeReleasedFileMarker(f); err != nil {
+		f.mu.Unlock()
+		return err
+	}
+	if err := me.removeFileStorage(f); err != nil {
+		cleanupErr := me.removeReleasedFileStorage(f)
+		f.mu.Unlock()
+		return errors.Join(err, cleanupErr)
+	}
 	f.released = true
-	if err := os.MkdirAll(filepath.Dir(f.releasedFilePath()), dirPerm); err != nil {
-		f.mu.Unlock()
-		return err
-	}
-	marker, err := os.OpenFile(f.releasedFilePath(), os.O_CREATE|os.O_TRUNC|os.O_WRONLY, filePerm)
-	if err != nil {
-		f.mu.Unlock()
-		return err
-	}
-	err = marker.Close()
 	f.mu.Unlock()
-	if err != nil {
-		return err
-	}
 	for pieceIndex := f.beginPieceIndex(); pieceIndex < f.endPieceIndex(); pieceIndex++ {
 		if err := me.setPieceCompletion(pieceIndex, true); err != nil {
 			return fmt.Errorf("setting released file piece %d complete: %w", pieceIndex, err)
@@ -78,21 +81,81 @@ func (me *fileTorrentImpl) markFileReleased(fileIndex int) error {
 	return nil
 }
 
+func (me *fileTorrentImpl) markFileDiscarded(fileIndex int) error {
+	if err := me.validateFileIndex(fileIndex); err != nil {
+		return err
+	}
+	f := me.file(fileIndex)
+	f.mu.Lock()
+	f.released = false
+	err := errors.Join(
+		me.removeFileStorage(f),
+		me.removeReleasedFileStorage(f),
+	)
+	f.mu.Unlock()
+	if err != nil {
+		return fmt.Errorf("discarding file storage: %w", err)
+	}
+	for pieceIndex := f.beginPieceIndex(); pieceIndex < f.endPieceIndex(); pieceIndex++ {
+		if err := me.setPieceCompletion(pieceIndex, false); err != nil {
+			return fmt.Errorf("setting discarded file piece %d incomplete: %w", pieceIndex, err)
+		}
+	}
+	return nil
+}
+
+func (me *fileTorrentImpl) writeReleasedFileMarker(f file) error {
+	if err := os.MkdirAll(filepath.Dir(f.releasedFilePath()), dirPerm); err != nil {
+		return err
+	}
+	marker, err := os.OpenFile(f.releasedFilePath(), os.O_CREATE|os.O_TRUNC|os.O_WRONLY, filePerm)
+	if err != nil {
+		return err
+	}
+	return marker.Close()
+}
+
+func (me *fileTorrentImpl) removeReleasedFileStorage(f file) error {
+	var err error
+	err = errors.Join(err, me.removeFilePath(f.releasedFilePath()))
+	for pieceIndex := f.beginPieceIndex(); pieceIndex < f.endPieceIndex(); pieceIndex++ {
+		err = errors.Join(err, me.removeFilePath(f.releasedPiecePath(pieceIndex)))
+	}
+	return err
+}
+
+func (me *fileTorrentImpl) removeFileStorage(f file) (err error) {
+	err = errors.Join(
+		me.removeFilePath(f.safeOsPath),
+		me.removeFilePath(f.partFilePath()),
+	)
+	if err != nil {
+		return fmt.Errorf("removing released file storage: %w", err)
+	}
+	return nil
+}
+
+func (me *fileTorrentImpl) removeFilePath(path string) error {
+	err := me.io.remove(path)
+	if err == nil || errors.Is(err, fs.ErrNotExist) {
+		return nil
+	}
+	if errors.Is(err, fs.ErrPermission) {
+		if chmodErr := os.Chmod(path, filePerm); chmodErr != nil {
+			return fmt.Errorf("preparing %q for removal: %w", path, chmodErr)
+		}
+		err = me.io.remove(path)
+		if err == nil || errors.Is(err, fs.ErrNotExist) {
+			return nil
+		}
+	}
+	return fmt.Errorf("removing %q: %w", path, err)
+}
+
 func (me *fileTorrentImpl) writeReleasedBoundaryPieces(fileIndex int, f file) error {
 	for pieceIndex := f.beginPieceIndex(); pieceIndex < f.endPieceIndex(); pieceIndex++ {
-		piece := me.info.Piece(pieceIndex)
-		pieceExtent := segments.Extent{Start: piece.Offset(), Length: piece.Length()}
-		var fileExtent segments.Extent
-		fileSegments := 0
-		foundFile := false
-		for i, extent := range me.segmentLocater.LocateIter(pieceExtent) {
-			fileSegments++
-			if i == fileIndex {
-				fileExtent = extent
-				foundFile = true
-			}
-		}
-		if !foundFile || fileSegments <= 1 {
+		fileExtent, ok := me.fileExtentInBoundaryPiece(fileIndex, pieceIndex)
+		if !ok {
 			continue
 		}
 		if err := me.writeReleasedBoundaryPiece(f, pieceIndex, fileExtent); err != nil {
@@ -100,6 +163,32 @@ func (me *fileTorrentImpl) writeReleasedBoundaryPieces(fileIndex int, f file) er
 		}
 	}
 	return nil
+}
+
+func (me *fileTorrentImpl) fileExtentInBoundaryPiece(fileIndex, pieceIndex int) (segments.Extent, bool) {
+	fileExtent, ok, fileSegments := me.fileExtentInPieceWithSegmentCount(fileIndex, pieceIndex)
+	return fileExtent, ok && fileSegments > 1
+}
+
+func (me *fileTorrentImpl) fileExtentInPiece(fileIndex, pieceIndex int) (segments.Extent, bool) {
+	fileExtent, ok, _ := me.fileExtentInPieceWithSegmentCount(fileIndex, pieceIndex)
+	return fileExtent, ok
+}
+
+func (me *fileTorrentImpl) fileExtentInPieceWithSegmentCount(
+	fileIndex int,
+	pieceIndex int,
+) (fileExtent segments.Extent, foundFile bool, fileSegments int) {
+	piece := me.info.Piece(pieceIndex)
+	pieceExtent := segments.Extent{Start: piece.Offset(), Length: piece.Length()}
+	for i, extent := range me.segmentLocater.LocateIter(pieceExtent) {
+		fileSegments++
+		if i == fileIndex {
+			fileExtent = extent
+			foundFile = true
+		}
+	}
+	return
 }
 
 func (me *fileTorrentImpl) writeReleasedBoundaryPiece(f file, pieceIndex int, extent segments.Extent) error {
@@ -237,6 +326,11 @@ func (me *fileTorrentImpl) loadReleasedFileMarkers() {
 // Open file for reading.
 func (me *fileTorrentImpl) openSharedFile(file file) (f sharableReader, err error) {
 	file.mu.RLock()
+	if file.released {
+		file.mu.RUnlock()
+		err = ErrFileReleased
+		return
+	}
 	// Fine to open once under each name on a unix system. We could make the shared file keys more
 	// constrained, but it shouldn't matter. TODO: Ensure at most one of the names exist.
 	if me.partFiles() {
@@ -252,6 +346,11 @@ func (me *fileTorrentImpl) openSharedFile(file file) (f sharableReader, err erro
 // Open file for reading. Not a shared handle if that matters.
 func (me *fileTorrentImpl) openFile(file file) (f fileReader, err error) {
 	file.mu.RLock()
+	if file.released {
+		file.mu.RUnlock()
+		err = ErrFileReleased
+		return
+	}
 	// Fine to open once under each name on a unix system. We could make the shared file keys more
 	// constrained, but it shouldn't matter. TODO: Ensure at most one of the names exist.
 	if me.partFiles() {
@@ -267,5 +366,12 @@ func (me *fileTorrentImpl) openFile(file file) (f fileReader, err error) {
 func (me *fileTorrentImpl) openForWrite(file file) (_ fileWriter, err error) {
 	// It might be possible to have a writable handle shared files cache if we need it.
 	me.logger().Debug("openForWrite", "file.safeOsPath", file.safeOsPath)
-	return me.io.openForWrite(me.pathForWrite(&file), file.FileInfo.Length)
+	file.mu.RLock()
+	if file.released {
+		file.mu.RUnlock()
+		return nil, ErrFileReleased
+	}
+	path := me.pathForWrite(&file)
+	file.mu.RUnlock()
+	return me.io.openForWrite(path, file.FileInfo.Length)
 }